requirement.txt
h5py==2.9.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-rectified-adam==0.12.0
matplotlib==3.1.1
numba==0.45.1
numpy==1.16.4
opencv-python==4.1.1.26
pandas==0.25.2
Pillow==6.1.0
pycuda==2016.1
python-apt==1.1.0b1+ubuntu0.16.4.5
python-dateutil==2.8.0
python-debian==0.1.27
python-systemd==231
pytools==2015.1.6
scikit-learn==0.21.3
scipy==1.3.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-plot==0.3.2
tensorrt==6.0.1.5

